fix(phpstan): clear 4 strict-rule errors slipping past baseline trim
- EmailChangeController.php:246/259 — (int) (CONST / 24) where the
  expression is already int; drop the cast.
- tests/Api/MediaObjectTest.php:32 — short ternary on glob() result;
  expand to long ternary with false-check.
- tests/Api/PushSubscriptionTest.php:74 — (string) on
  Response::getContent() which already returns string.

+++ api/src/Controller/EmailChangeController.php
@@ -243,7 +243,7 @@ class EmailChangeController extends AbstractController
                 $changeRequest->getNewEmail(),
                 $revertUrl,
                 $resetUrl,
-                (int) (self::REVERT_TOKEN_TTL_HOURS / 24),
+                self::REVERT_TOKEN_TTL_HOURS / 24,
             ))
             ->html(sprintf(
                 '<p>Hi,</p>'
@@ -256,7 +256,7 @@ class EmailChangeController extends AbstractController
                 htmlspecialchars($changeRequest->getNewEmail()),
                 htmlspecialchars($revertUrl),
                 htmlspecialchars($resetUrl),
-                (int) (self::REVERT_TOKEN_TTL_HOURS / 24),
+                self::REVERT_TOKEN_TTL_HOURS / 24,
             ));
 
         $this->mailer->send($email);
